Split block explosion event out
