+++ src/main/kotlin/one/oktw/galaxy/block/util/CustomBlockUtil.kt
@@ -81,10 +81,14 @@ object CustomBlockUtil {
         return true
     }
 
-    fun positionIsBlock(world: ServerWorld, blockPos: BlockPos, type: BlockType): Boolean {
+    fun positionMatchesCustomBlock(world: ServerWorld, blockPos: BlockPos, type: BlockType): Boolean {
         val entity = getCustomBlockEntity(world, blockPos) ?: return false
         val blockType = getTypeFromCustomBlockEntity(entity) ?: return false
         return blockType == type
+    }
+
+    fun positionIsAnyCustomBlock(world: ServerWorld, blockPos: BlockPos): Boolean {
+        return getCustomBlockEntity(world, blockPos) != null
     }
 
     fun getCustomBlockEntity(world: ServerWorld, blockPos: BlockPos): Entity? {
